+++ src/main/java/DealerCards.java
@@ -1,59 +1,14 @@
-import java.util.Collections;
 import java.util.List;
 
-public class DealerCards {
+public class DealerCards extends GamerCards {
 
-    private static final int BUST_THRESHOLD = 21;
-    private static final int CONDITION_DECIDING_A_SCORE = 10;
-    private static final int A_SCORE_GAP = 10;
-    public static final int DEALER_HIT_THRESHOLD = 16;
-
-    private final List<Card> cards;
+    private static final int DEALER_HIT_THRESHOLD = 16;
 
     public DealerCards(List<Card> cards) {
-        this.cards = cards;
-    }
-
-    public void addCard(Card card) {
-        cards.add(card);
-    }
-
-    public int calculateScore() {
-        int totalScore = sumExceptAceCards();
-        totalScore += sumAceCards(totalScore);
-        return totalScore;
-    }
-
-    private int sumExceptAceCards() {
-        return cards.stream()
-                .filter(Card::isNotAce)
-                .mapToInt(Card::score)
-                .sum();
-    }
-
-    private int sumAceCards(int score) {
-        return cards.stream()
-                .filter(Card::isAce)
-                .mapToInt(card -> decideAceScore(score))
-                .sum();
-    }
-
-    private int decideAceScore(int totalScore) {
-        if (totalScore <= CONDITION_DECIDING_A_SCORE) {
-            return CardNumber.A.getScore() + A_SCORE_GAP;
-        }
-        return CardNumber.A.getScore();
-    }
-
-    public boolean hasScoreUnderBustThreshold() {
-        return calculateScore() <= BUST_THRESHOLD;
+        super(cards);
     }
 
     public boolean hasScoreUnderHitThreshold() {
         return calculateScore() <= DEALER_HIT_THRESHOLD;
-    }
-
-    public List<Card> getCards() {
-        return Collections.unmodifiableList(cards);
     }
 }
